requirements.txt
fastapi==0.95.1
uvicorn[standard]
SQLAlchemy==2.0.12
python-dotenv==1.0.0
python-multipart==0.0.6
autopep8==2.0.2
pydantic[email]
PyMySQL==1.0.3
oracledb==1.3.1
